Implemented primitive REST query for [All / Review / Recipe type post entries]

+++ ConvApp/ConvApp/Models/User.cs
@@ -4,8 +4,7 @@ namespace ConvApp.Models
 {
     public class User
     {
-
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string ProfileImage { get; set; }
         public List<Posting> CreatedPostings { get; set; }
